# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EmailLifecycleRequestDto": {
        "properties": {
          "email": {
            "example": "ban@example.com",
            "format": "email",
            "maxLength": 320,
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "LoginDto": {
        "properties": {
          "email": {
            "example": "ban@example.com",
            "format": "email",
            "maxLength": 320,
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "TokenConfirmationDto": {
        "properties": {
          "token": {
            "maxLength": 128,
            "minLength": 32,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "token"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "bep_nho_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "HTTP contract for the Bếp Nhớ alpha API.",
    "title": "Bếp Nhớ API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/email-verification/confirm": {
      "post": {
        "operationId": "AuthController_confirmEmailVerification",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenConfirmationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Email verified."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "BAD_REQUEST"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "TOO_MANY_REQUESTS"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Consume a single-use email verification token.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/email-verification/request": {
      "post": {
        "operationId": "AuthController_requestEmailVerification",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLifecycleRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The request was accepted regardless of account existence."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "BAD_REQUEST"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "TOO_MANY_REQUESTS"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Request a single-use verification email with an enumeration-safe response.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authenticated."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "BAD_REQUEST"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "FORBIDDEN"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "TOO_MANY_REQUESTS"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "SERVICE_UNAVAILABLE"
          }
        },
        "summary": "Authenticate and establish a cookie session.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session cookie cleared."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "FORBIDDEN"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Revoke the current cookie session when present.",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "MeController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user and Taste profile."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "example": "VALIDATION_ERROR",
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "FORBIDDEN"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Return the current authenticated user context.",
        "tags": [
          "Identity"
        ]
      }
    },
...
  },
  "servers": [],
...
}